gui: optimized repetition point rendering

+++ src/odemis/gui/comp/canvas.py
@@ -1031,8 +1031,8 @@ def world_to_buffer_pos(world_pos, world_buffer_center, scale, offset=None):
     :param world_buffer_center: the center of the buffer in world coordinates
     :param scale: how much zoomed is the buffer compared to the world
     """
-    buff_pos = (round((world_pos[0] - world_buffer_center[0]) * scale),
-                round((world_pos[1] - world_buffer_center[1]) * scale))
+    buff_pos = (int(round((world_pos[0] - world_buffer_center[0]) * scale)),
+                int(round((world_pos[1] - world_buffer_center[1]) * scale)))
     if offset:
         return (buff_pos[0] + offset[0], buff_pos[1] + offset[1])
     else:
